+++ fractal_server/app/db/__init__.py
@@ -54,7 +54,6 @@ class DB:
             cls._engine_async,
             class_=AsyncSession,
             expire_on_commit=False,
-            future=True,
         )
 
     @classmethod
@@ -68,12 +67,7 @@ class DB:
             pool_pre_ping=True,
         )
 
-        cls._sync_session_maker = sessionmaker(
-            bind=cls._engine_sync,
-            autocommit=False,
-            autoflush=False,
-            future=True,
-        )
+        cls._sync_session_maker = sessionmaker(cls._engine_sync)
 
     @classmethod
     async def get_async_db(cls) -> AsyncGenerator[AsyncSession, None]: